Fixed wrong type for some draws

+++ src/core/gfx/graphics.py
@@ -218,10 +218,10 @@ class Graphics:
         if ln < 2:
             return
 
-        filled_circle(surface, points[0][0], points[0][1], int(stroke_weight / 2), stroke_color)
+        filled_circle(surface, int(points[0][0]), int(points[0][1]), int(stroke_weight / 2), stroke_color)
         for i in range(1, ln):
             Graphics.line(surface, points[i - 1], points[i], stroke_weight, stroke_color)
-            filled_circle(surface, points[i][0], points[i][1], int(stroke_weight / 2), stroke_color)
+            filled_circle(surface, int(points[i][0]), int(points[i][1]), int(stroke_weight / 2), stroke_color)
 
     @staticmethod
     def arrow_line(
